main print incfile

+++ vm/inc/corewar_vm.h
@@ -74,9 +74,9 @@ typedef struct 			s_general
 	size_t 				cnt_live;
 	size_t 				cycles_to_die;
 	size_t				num_checks;
-	int					dump_cycle; //-dump option
-    int					pl_nbr; //number of player's
-    int					verb_nbr; //verbosity option
+	size_t				dump_cycle; //-dump option
+    size_t				pl_nbr; //number of player's
+    uint8_t				verb_nbr; //verbosity option
 }						t_general;
